move migration script into dedicated folder
